Moved  Parquet Sink test to correct directory.
